backend/requirements.txt
Flask==3.0.3
flask-pymongo==2.3.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
dnspython==2.6.1
python-dotenv==1.0.1
Flask-Cors==4.0.0
pytz==2024.1
Flask-SQLAlchemy
Flask-Migrate
email-validator
requests
yfinance
pandas
nsetools
pymongo==4.8.0

# ------------------------
# Date & Time Utilities
# ------------------------
pytz==2024.1

# ------------------------
# Background Scheduler (for automatic email reminders)
# ------------------------
APScheduler==3.10.4

# ------------------------
# Data Validation / Safety (optional but useful)
# ------------------------
Werkzeug==3.0.3
itsdangerous==2.2.0

# ------------------------
# Development / Debugging
# ------------------------
dnspython==2.6.1